feat: 实现雪花ID生成器和BigDecimal金额精度修复
- 添加雪花ID生成器工具类和JPA配置
- 修复所有实体类的金额字段精度问题
- 更新数据库DDL脚本
- 删除过时的数据库对齐文档
- 添加详细的修复总结文档

+++ src/main/java/org/ivan/artshow/module/course/pojo/dto/CourseDTO.java
@@ -1,5 +1,7 @@
 package org.ivan.artshow.module.course.pojo.dto;
 
+import java.math.BigDecimal;
+
 /**
  * CourseDTO - 数据传输对象
  *
@@ -13,7 +15,7 @@ public record CourseDTO(
         Integer instructorId,
         String title,
         String coverImageUrl,
-        Double price,
+        BigDecimal price,
         String type,
         Integer studentCount,
         String description
@@ -22,7 +24,7 @@ public record CourseDTO(
     public Integer getInstructorId() { return instructorId; }
     public String getTitle() { return title; }
     public String getCoverImageUrl() { return coverImageUrl; }
-    public Double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
     public String getType() { return type; }
     public Integer getStudentCount() { return studentCount; }
     public String getDescription() { return description; }
